+++ gulpfile.js
@@ -1,6 +1,6 @@
 "use strict";
 
-const gulp = require( "@gulpx/api" );
+const gulp = require( "@futagoza/gulpx" );
 const bump = require( "@futagoza/gulp/bump" );
 const eslint = require( "gulp-eslint" );
 const publish = require( "@futagoza/gulp/publish" );
@@ -10,8 +10,7 @@ const pump = require( "pump" );
 gulp.task( "lint", () => pump(
 
     gulp.src( [
-        "gulpx/cli/bin/gulpx",
-        "gulpx/**/*.js",
+        "packages/gulpx/bin/gulpx",
         "packages/**/*.js",
         "plugins/**/*.js",
         ".eslintrc.js",
@@ -20,15 +19,6 @@ gulp.task( "lint", () => pump(
     eslint( { dotfiles: true } ),
     eslint.format(),
     eslint.failAfterError()
-
-) );
-
-// Bump the "version" field of every `package.json` in the gulpx directory
-gulp.task( "bump:gulpx", () => pump(
-
-    gulp.src( "gulpx/**/package.json" ),
-    bump(),
-    gulp.dest( "gulpx/" )
 
 ) );
 
@@ -51,12 +41,11 @@ gulp.task( "bump:plugins", () => pump(
 ) );
 
 // Bump everything (good idea?)
-gulp.task( "bump", gulp.series( "bump:gulpx", "bump:packages", "bump:plugins" ) );
+gulp.task( "bump", gulp.series( "bump:packages", "bump:plugins" ) );
 
 // Publish all the packages in this monorepo (versions should be synced before-hand)
 gulp.task( "publish", () => pump(
     gulp.src( [
-        "gulpx/*",
         "packages/*",
         "plugins/*",
     ] ),