+++ CHANGELOG.md
@@ -8,10 +8,9 @@
 * Set `@futagoza/gulp` to use `@futagoza/child-process`
 * Set `@futagoza/node-run` to use `@futagoza/child-process`
 * Set `@futagoza/publish-package` to use `@futagoza/child-process`
-* Added `@gulpx/api`
 * `@futagoza/pump` now accepts a single item
-* Added `@gulpx/cli`
 * `@futagoza/pump` only accepts a function, promise or stream
+* Added `@futagoza/gulpx`, a custom variant of the Gulp client and the Gulp CLI
 
 ## [v0.5.1](https://github.com/futagoza/gulp/compare/v0.5.0...v0.5.1) _(2018-09-26)_